backend/requirements.txt
# Core Dependencies
fastapi
uvicorn
pipecat-ai[daily,openai,deepgram,cartesia,langchain]
python-dotenv
pydantic
langchain
langchain-core
langchain-openai
langchain-google-genai
firebase-admin
gunicorn
